bot/requirements.txt
# Discord Bot - Standalone Dependencies
# Install with: pip install -r bot/requirements.txt

discord.py>=2.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
